+++ src/commands/images/index.ts
@@ -20,6 +20,10 @@ export const options: CommandOptions = [
 		flags: '-i, --images <images to deploy>',
 		description: 'define images to deploy',
 		defaultValue: '*'
+	},
+	{
+		flags: '-p, --push',
+		description: 'Push files to Amazon'
 	}
 ];
 
@@ -37,6 +41,7 @@ export const action = async (params: Command) => {
 	execs.display('Creating build.');
 
 	const sw_version = Date.now().toString(16);
+
 	for (const zone in zones) {
 		flivity.amazon.region = zone;
 
@@ -87,14 +92,16 @@ export const action = async (params: Command) => {
 		}
 	}
 
-	execs.display('\nUploading images to Elastic Container Registry.');
-	for (const region_name in server_images) {
-		flivity.amazon.region = region_name;
+	if (currentOptions.push) {
+		execs.display('\nUploading images to Elastic Container Registry.');
+		for (const region_name in server_images) {
+			flivity.amazon.region = region_name;
 
-		for (const server_image of server_images[region_name]) {
-			execs.display(`=> Uploading '${server_image}'.`, true);
-			execs.execute(`docker push ${server_image}:latest`);
-			execs.execute(`docker image rm ${server_image}`);
+			for (const server_image of server_images[region_name]) {
+				execs.display(`=> Uploading '${server_image}'.`, true);
+				execs.execute(`docker push ${server_image}:latest`);
+				execs.execute(`docker image rm ${server_image}`);
+			}
 		}
 	}
 };
